Implementing Factory Pattern Part3
Replacing all the convertToDTO methods by "DTO_Factory.create(typeOfObejct)". This delegate all the complexity of creating the object to the factory.
Note: In previous steps, I forget to create the overload for the create method in the factory corresponding to "convertCustomerToDTO". This was fixed now.
All the JUNIT test still passing