Implementing the Decorator Pattern making OrderDTO visitable
making the OrderDTO visitable will allow to able to implement the visitable interface (not created yet)
Note: to able to enhance the capabilities of OrderDTO without modifying the class itself, the OrderDTOVisitable will have a reference to the Order TO that will be initialised in its constructor. However, we need to implement all public methods from OrderDTO in OrderDTOVisitable. Inside this method will be calling to the corresponding method in the OrderDTO reference that we have.